+++ src/compas_libigl/mapping.py
@@ -26,7 +26,7 @@ from compas_libigl._types_std import VectorInt  # noqa: F401
 from compas_libigl._types_std import VectorVectorInt  # noqa: F401
 
 
-def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
+def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, simplify_borders=True, fixed_vertices=None, tolerance=1e-6):
     """
     Map a 2D pattern mesh onto a 3D target.
 
@@ -36,10 +36,18 @@ def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
         A tuple of (vertices, faces) representing the target mesh.
     pattern_mesh : tuple[list[list[float]], list[list[int]]]
         A tuple of (vertices, faces) representing the pattern mesh.
-    clip_boundaries : bool
+    clip_boundaries : bool, optional
         Whether to clip the pattern mesh to the boundaries of the target mesh.
-    tolerance : float
+        Default is True.
+    simplify_borders : bool, optional
+        Whether to simplify the border of the pattern mesh.
+        Default is True.
+    fixed_vertices : list[list[float]], optional
+        A list of fixed points on the target mesh.
+        Default is None.
+    tolerance : float, optional
         The tolerance for point comparison, to remove duplicates.
+        Default is 1e-6.
 
     Returns
     -------
@@ -61,16 +69,30 @@ def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
     f_numpy = np.array(f, dtype=np.int32)
     pattern_v_numpy = np.array(pv, dtype=np.float64)
 
+    # Handle fixed_vertices - provide empty array if None
+
+    fixed_vertices_vectorint = VectorInt()
+    if fixed_vertices is None:
+        fixed_vertices_vectorint = VectorInt()
+    else:
+        fixed_vertices_vectorint = VectorInt(fixed_vertices)
+
+    # Convert pattern_f from Python list to VectorVectorInt which is expected by C++ code
+
+    pattern_f_vec = VectorVectorInt()
+    for face in pf:
+        pattern_f_vec.append(face)
+
     # Perform the mapping
     pv_numpy_copy, pf_numpy_cleaned, p_normals, pattern_is_boundary, pattern_groups = _mapping.map_mesh_with_automatic_parameterization(
-        v_numpy, f_numpy, pattern_v_numpy, pf, clip_boundaries, tolerance
+        v_numpy, f_numpy, pattern_v_numpy, pattern_f_vec, clip_boundaries, simplify_borders, fixed_vertices_vectorint, tolerance
     )
 
     # Return the result as a tuple
     return pv_numpy_copy, pf_numpy_cleaned, p_normals, pattern_is_boundary, pattern_groups
 
 
-def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, pattern_u=16, pattern_v=16):
+def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, pattern_u=16, pattern_v=16, simplify_borders=True, fixed_vertices=None):
     """
     Map a 2D pattern mesh onto a 3D target.
 
@@ -100,14 +122,24 @@ def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, patter
 
     mesh : compas.datastructures.Mesh
         The target mesh.
-    clip_boundaries : bool
+    clip_boundaries : bool, optional
         Whether to clip the pattern mesh to the boundaries of the target mesh.
-    tolerance : float
+        Default is True.
+    tolerance : float, optional
         The tolerance for point comparison, to remove duplicates.
-    pattern_u : int
+        Default is 1e-6.
+    pattern_u : int, optional
         The number of pattern vertices in the u direction.
-    pattern_v : int
+        Default is 16.
+    pattern_v : int, optional
         The number of pattern vertices in the v direction.
+        Default is 16.
+    simplify_borders : bool, optional
+        Whether to simplify the border of the pattern mesh.
+        Default is True.
+    fixed_vertices : list[list[float]], optional
+        A list of fixed points on the target mesh.
+        Default is None.
 
     Returns
     -------
@@ -165,6 +197,8 @@ def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, patter
     pf = tessagon_mesh["face_list"]
 
     v, f = mesh.to_vertices_and_faces()
-    mapped_vertices, mapped_faces, mapped_normals, mapped_is_boundary, mapped_groups = map_mesh((v, f), (pv, pf), clip_boundaries=clip_boundaries, tolerance=tolerance)
+    mapped_vertices, mapped_faces, mapped_normals, mapped_is_boundary, mapped_groups = map_mesh(
+        (v, f), (pv, pf), clip_boundaries=clip_boundaries, simplify_borders=simplify_borders, fixed_vertices=fixed_vertices, tolerance=tolerance
+    )
 
     return Mesh.from_vertices_and_faces(mapped_vertices, mapped_faces)