+++ CHANGELOG.md
@@ -18,8 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added simplified borders to map_mesh function.
+- Added fixed points to map_mesh function.
+
 ### Changed
 
+- Fixed point welding bug in map_mesh function.
+- Changed example_isolines to work with the new compas_viewer.
 - GA for cibuildwheel, bring back the tessagon test.
 
 ### Removed